feat: move interactive branch selection to new log short display
